clients/js: refactor shared code
